Users and Courses Collections
User collection created and added to whenever a user that has not logged in before opens the home page (once they log in). The courses collection can be added to using the upload page and uploading notes a class that has not been defined already. It definitely needs some work, but the basic functionality is there.